third_party/WebKit/LayoutTests/fast/dom/shadow/shadowdom-for-input-button-click-expected.txt

CONSOLE WARNING: <shadow> doesn't work for INPUT element host.

Make sure that mouse click on an input button with shadow nodes works.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Click on the padding area of the button:
PASS clickCount is 1
Click on the margin area of a shadow node:
PASS clickCount is 1
Click on the middle of a shadow node:
PASS clickCount is 1
Click on the original value text:
PASS clickCount is 1
PASS successfullyParsed is true

TEST COMPLETE
